Adding single-element api for subnetwork
