+++ tests/utils/test_networkx_utils.py
@@ -13,8 +13,6 @@ from src.utils.networkx_utils import (
     yield_edges,
     yield_nodes,
     yield_edges_with_nodes,
-    is_dict_subset,
-    is_subarray,
 )
 
 
@@ -423,23 +421,6 @@ def test_yield_edges_with_nodes():
     ]
 
 
-def test_is_dict_subset():
-    assert is_dict_subset(dict(), dict())  # empty, empty
-    assert is_dict_subset(dict(), {"A": 1})  # empty, non-empty
-    assert is_dict_subset({"A": 1}, {"A": 1})  # exact match, single
-    assert is_dict_subset({"A": 1, "B": 2}, {"A": 1, "B": 2})  # exact match, multiple
-    assert is_dict_subset({"A": 1}, {"A": 1, "B": 2})  # subset single
-    assert is_dict_subset({"A": 1, "B": 2}, {"A": 1, "B": 2, "C": 3})  # subset multiple
-
-    assert not is_dict_subset({"A": 1}, dict())  # non-empty,
-    assert not is_dict_subset({"A": 1}, {"B": 2})  # disjoint
-    assert not is_dict_subset({"A": 1,}, {"A": 2})  # larger set wrong value
-    assert not is_dict_subset(
-        {"A": 1, "B": 2}, {"A": 1, "B": 1}
-    )  # larger set wrong value multiple
-    assert not is_dict_subset({"A": 1, "B": 2}, {"A": 1})  # larger set missing element
-
-
 def test_get_edges_by_division():
     test_sdl = """
     # edges
@@ -495,22 +476,3 @@ def test_get_edges_by_division():
         ("A", "B", 0),
         ("C", "D", 0),
     ]
-
-
-def test_is_subarray():
-    assert is_subarray([1], [1])  # exactly equal
-    assert is_subarray([1, 2], [1, 2])  # exactly equal, multiple
-    assert is_subarray([1], [1, 2])  # left
-    assert is_subarray([1, 2], [1, 2, 3])  # left, multiple
-    assert is_subarray([1], [2, 1])  # right
-    assert is_subarray([1, 0], [2, 1, 0])  # right, multiple
-    assert is_subarray([1], [3, 1, 2])  # middle
-    assert is_subarray([1, 0], [2, 1, 0, -1])  # middle, multiple
-
-    assert not is_subarray([], [])  # empty empty
-    assert not is_subarray([], [1])  # empty, nonempty
-    assert not is_subarray([1], [])  # nonempty, empty
-    assert not is_subarray([1], [2])  # not present single
-    assert not is_subarray([1, 2], [1, 3])  # not present multiple
-    assert not is_subarray([1, 2], [1, 0, 2])  # present, but not sequential
-    assert not is_subarray([1, 2], [2, 1])  # present, but reversed
