+++ todo.txt
@@ -1,11 +1,8 @@
 TODO
-Allow multiple divisions in subnetwork divisions api
-Allow single element in subnetwork api
 Write tests for new drama_network api
 Write tests for networkx converter division data
 Write function for combining edges together by adding weights, user-provided function for comparing data
 Incorporate support for additional, user-defined sections
-Think about an empty DramaNetwork
 Refactor test_networkx_utils to clean up duplicate test data
 Error when some edges are underneath divisions and others aren't
 Write Sphinx-compatible docstrings for all the public functions
@@ -38,3 +35,6 @@ Add shallow copy graph subsetting by node data_value X
 Add shallow copy graph subsetting by by edge data_value X
 Add drama network api for subgraph X
 Add division data to drama_network api X
+Allow multiple divisions in subnetwork divisions api X
+Allow single element in subnetwork api X
+Think about an empty DramaNetwork X
